Fix modal issues by creating a Component to wrap the example
